Rename the package name by avoiding duplications
