+++ runner/components/io.siddhi.distribution.core/src/gen/java/io/siddhi/distribution/core/api/ApiResponseMessageWithCode.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package io.siddhi.distribution.core.api.api;
+package io.siddhi.distribution.core.api;
 
 import javax.xml.bind.annotation.XmlTransient;
