+++ runner/components/io.siddhi.distribution.core/src/gen/java/io/siddhi/distribution/core/api/RFC3339DateFormat.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package io.siddhi.distribution.core.api.api;
+package io.siddhi.distribution.core.api;
 
 import com.fasterxml.jackson.databind.util.ISO8601DateFormat;
 import com.fasterxml.jackson.databind.util.ISO8601Utils;
