+++ runner/components/io.siddhi.distribution.core/src/gen/java/io/siddhi/distribution/core/api/RunTimeApiService.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2017, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
+ * Copyright (c) 2018, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,19 +14,18 @@
  * limitations under the License.
  */
 
-package io.siddhi.distribution.core.api.api;
+package io.siddhi.distribution.core.api;
+
+import org.wso2.msf4j.Request;
+
+import javax.ws.rs.core.Response;
 
 /**
  * Auto generated class from Swagger to MSF4J.
  */
 
 @javax.annotation.Generated(value = "io.swagger.codegen.languages.JavaMSF4JServerCodegen",
-        date = "2017-03-15T08:56:59.657Z")
-public class NotFoundException extends ApiException {
-    private int code;
-
-    public NotFoundException(int code, String msg) {
-        super(code, msg);
-        this.code = code;
-    }
+        date = "2018-02-20T10:49:25.745Z")
+public abstract class RunTimeApiService {
+    public abstract Response getRunTime(Request request) throws NotFoundException;
 }
