+++ runner/components/io.siddhi.distribution.core/src/gen/java/io/siddhi/distribution/core/api/StringUtil.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package io.siddhi.distribution.core.api.api;
+package io.siddhi.distribution.core.api;
 
 /**
  * Auto generated class from Swagger to MSF4J.
