+++ runner/components/io.siddhi.distribution.core/src/gen/java/io/siddhi/distribution/core/model/InlineResponse200.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package io.siddhi.distribution.core.model.model;
+package io.siddhi.distribution.core.model;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import io.swagger.annotations.ApiModelProperty;
@@ -22,75 +22,53 @@ import io.swagger.annotations.ApiModelProperty;
 import java.util.Objects;
 
 /**
- * Error.
+ * InlineResponse200.
  */
 @javax.annotation.Generated(value = "io.swagger.codegen.languages.JavaMSF4JServerCodegen",
         date = "2017-05-31T15:43:24.557Z")
-public class Error {
-    @JsonProperty("code")
-    private Integer code = null;
+public class InlineResponse200 {
+    @JsonProperty("name")
+    private String name = null;
 
-    @JsonProperty("status")
-    private String status = null;
+    @JsonProperty("query")
+    private String query = null;
 
-    @JsonProperty("message")
-    private String message = null;
-
-    public Error code(Integer code) {
-        this.code = code;
+    public InlineResponse200 name(String name) {
+        this.name = name;
         return this;
     }
 
     /**
-     * Get code.
+     * Siddhi App Name.
      *
-     * @return code
+     * @return name
      **/
-    @ApiModelProperty(value = "")
-    public Integer getCode() {
-        return code;
+    @ApiModelProperty(value = "Siddhi App Name")
+    public String getName() {
+        return name;
     }
 
-    public void setCode(Integer code) {
-        this.code = code;
+    public void setName(String name) {
+        this.name = name;
     }
 
-    public Error status(String status) {
-        this.status = status;
+    public InlineResponse200 query(String query) {
+        this.query = query;
         return this;
     }
 
     /**
-     * Get status.
+     * Siddhi Query.
      *
-     * @return status
+     * @return query
      **/
-    @ApiModelProperty(value = "")
-    public String getStatus() {
-        return status;
+    @ApiModelProperty(value = "Siddhi Query")
+    public String getQuery() {
+        return query;
     }
 
-    public void setStatus(String status) {
-        this.status = status;
-    }
-
-    public Error message(String message) {
-        this.message = message;
-        return this;
-    }
-
-    /**
-     * Get message.
-     *
-     * @return message
-     **/
-    @ApiModelProperty(value = "")
-    public String getMessage() {
-        return message;
-    }
-
-    public void setMessage(String message) {
-        this.message = message;
+    public void setQuery(String query) {
+        this.query = query;
     }
 
 
@@ -102,25 +80,23 @@ public class Error {
         if (o == null || getClass() != o.getClass()) {
             return false;
         }
-        Error error = (Error) o;
-        return Objects.equals(this.code, error.code) &&
-                Objects.equals(this.status, error.status) &&
-                Objects.equals(this.message, error.message);
+        InlineResponse200 inlineResponse200 = (InlineResponse200) o;
+        return Objects.equals(this.name, inlineResponse200.name) &&
+                Objects.equals(this.query, inlineResponse200.query);
     }
 
     @Override
     public int hashCode() {
-        return Objects.hash(code, status, message);
+        return Objects.hash(name, query);
     }
 
     @Override
     public String toString() {
         StringBuilder sb = new StringBuilder();
-        sb.append("class Error {\n");
+        sb.append("class InlineResponse200 {\n");
 
-        sb.append("    code: ").append(toIndentedString(code)).append("\n");
-        sb.append("    status: ").append(toIndentedString(status)).append("\n");
-        sb.append("    message: ").append(toIndentedString(message)).append("\n");
+        sb.append("    name: ").append(toIndentedString(name)).append("\n");
+        sb.append("    query: ").append(toIndentedString(query)).append("\n");
         sb.append("}");
         return sb.toString();
     }
