+++ runner/components/io.siddhi.distribution.core/src/gen/java/io/siddhi/distribution/core/model/InlineResponse400.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package io.siddhi.distribution.core.model.model;
+package io.siddhi.distribution.core.model;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import io.swagger.annotations.ApiModelProperty;
@@ -22,11 +22,11 @@ import io.swagger.annotations.ApiModelProperty;
 import java.util.Objects;
 
 /**
- * Success.
+ * InlineResponse400.
  */
 @javax.annotation.Generated(value = "io.swagger.codegen.languages.JavaMSF4JServerCodegen",
         date = "2017-05-31T15:43:24.557Z")
-public class Success {
+public class InlineResponse400 {
     @JsonProperty("code")
     private Integer code = null;
 
@@ -36,7 +36,7 @@ public class Success {
     @JsonProperty("message")
     private String message = null;
 
-    public Success code(Integer code) {
+    public InlineResponse400 code(Integer code) {
         this.code = code;
         return this;
     }
@@ -55,7 +55,7 @@ public class Success {
         this.code = code;
     }
 
-    public Success status(String status) {
+    public InlineResponse400 status(String status) {
         this.status = status;
         return this;
     }
@@ -74,7 +74,7 @@ public class Success {
         this.status = status;
     }
 
-    public Success message(String message) {
+    public InlineResponse400 message(String message) {
         this.message = message;
         return this;
     }
@@ -102,10 +102,10 @@ public class Success {
         if (o == null || getClass() != o.getClass()) {
             return false;
         }
-        Success success = (Success) o;
-        return Objects.equals(this.code, success.code) &&
-                Objects.equals(this.status, success.status) &&
-                Objects.equals(this.message, success.message);
+        InlineResponse400 inlineResponse400 = (InlineResponse400) o;
+        return Objects.equals(this.code, inlineResponse400.code) &&
+                Objects.equals(this.status, inlineResponse400.status) &&
+                Objects.equals(this.message, inlineResponse400.message);
     }
 
     @Override
@@ -116,7 +116,7 @@ public class Success {
     @Override
     public String toString() {
         StringBuilder sb = new StringBuilder();
-        sb.append("class Success {\n");
+        sb.append("class InlineResponse400 {\n");
 
         sb.append("    code: ").append(toIndentedString(code)).append("\n");
         sb.append("    status: ").append(toIndentedString(status)).append("\n");
