+++ runner/components/io.siddhi.distribution.core/src/gen/java/io/siddhi/distribution/core/model/SiddhiAppMetrics.java
@@ -1,4 +1,4 @@
-package io.siddhi.distribution.core.model.model;
+package io.siddhi.distribution.core.model;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import io.swagger.annotations.ApiModelProperty;
