+++ runner/components/io.siddhi.distribution.core/src/gen/java/io/siddhi/distribution/core/model/SiddhiAppRevision.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package io.siddhi.distribution.core.model.model;
+package io.siddhi.distribution.core.model;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import io.swagger.annotations.ApiModelProperty;
@@ -26,33 +26,33 @@ import java.util.Objects;
  */
 @javax.annotation.Generated(value = "io.swagger.codegen.languages.JavaMSF4JServerCodegen",
         date = "2017-05-31T15:43:24.557Z")
-public class SiddhiAppContent {
+public class SiddhiAppRevision {
 
-    @JsonProperty("content")
-    private String content = null;
+    @JsonProperty("revision")
+    private String revision = null;
 
-    public SiddhiAppContent name(String name) {
+    public SiddhiAppRevision name(String name) {
         return this;
     }
 
 
-    public SiddhiAppContent content(String content) {
-        this.content = content;
+    public SiddhiAppRevision revision(String revision) {
+        this.revision = revision;
         return this;
     }
 
     /**
-     * Siddhi content.
+     * Siddhi revision.
      *
-     * @return content
+     * @return revision
      **/
-    @ApiModelProperty(value = "Siddhi content")
-    public String getcontent() {
-        return content;
+    @ApiModelProperty(value = "Siddhi revision")
+    public String getrevision() {
+        return revision;
     }
 
-    public void setcontent(String content) {
-        this.content = content;
+    public void setrevision(String revision) {
+        this.revision = revision;
     }
 
 
@@ -64,20 +64,20 @@ public class SiddhiAppContent {
         if (o == null || getClass() != o.getClass()) {
             return false;
         }
-        SiddhiAppContent siddhiAppContent = (SiddhiAppContent) o;
-        return Objects.equals(this.content, siddhiAppContent.content);
+        SiddhiAppRevision siddhiAppContent = (SiddhiAppRevision) o;
+        return Objects.equals(this.revision, siddhiAppContent.revision);
     }
 
     @Override
     public int hashCode() {
-        return Objects.hash(content);
+        return Objects.hash(revision);
     }
 
     @Override
     public String toString() {
         StringBuilder sb = new StringBuilder();
         sb.append("class SiddhiAppContent {\n");
-        sb.append("    content: ").append(toIndentedString(content)).append("\n");
+        sb.append("    revision: ").append(toIndentedString(revision)).append("\n");
         sb.append("}");
         return sb.toString();
     }
