+++ runner/components/io.siddhi.distribution.core/src/main/java/io/siddhi/distribution/core/factories/RunTimeApiServiceFactory.java
@@ -16,7 +16,7 @@
 
 package io.siddhi.distribution.core.factories;
 
-import io.siddhi.distribution.core.api.api.RunTimeApiService;
+import io.siddhi.distribution.core.api.RunTimeApiService;
 import io.siddhi.distribution.core.impl.RunTimeApiServiceImpl;
 
 /**
