+++ runner/components/io.siddhi.distribution.core/src/main/java/io/siddhi/distribution/core/factories/SiddhiAppsApiServiceFactory.java
@@ -16,7 +16,7 @@
 
 package io.siddhi.distribution.core.factories;
 
-import io.siddhi.distribution.core.api.api.SiddhiAppsApiService;
+import io.siddhi.distribution.core.api.SiddhiAppsApiService;
 import io.siddhi.distribution.core.impl.SiddhiAppsApiServiceImpl;
 
 /**
