+++ runner/components/io.siddhi.distribution.core/src/main/java/io/siddhi/distribution/core/impl/SiddhiAppsApiServiceImpl.java
@@ -17,10 +17,10 @@
 package io.siddhi.distribution.core.impl;
 
 import com.google.gson.Gson;
-import io.siddhi.distribution.core.api.api.ApiResponseMessage;
-import io.siddhi.distribution.core.api.api.ApiResponseMessageWithCode;
-import io.siddhi.distribution.core.api.api.NotFoundException;
-import io.siddhi.distribution.core.api.api.SiddhiAppsApiService;
+import io.siddhi.distribution.core.api.ApiResponseMessage;
+import io.siddhi.distribution.core.api.ApiResponseMessageWithCode;
+import io.siddhi.distribution.core.api.NotFoundException;
+import io.siddhi.distribution.core.api.SiddhiAppsApiService;
 import io.siddhi.distribution.core.impl.utils.Constants;
 import io.siddhi.distribution.core.internal.SiddhiAppData;
 import io.siddhi.distribution.core.internal.StreamProcessorDataHolder;
@@ -28,10 +28,10 @@ import io.siddhi.distribution.core.internal.beans.SiddhiAppElements;
 import io.siddhi.distribution.core.internal.exception.SiddhiAppConfigurationException;
 import io.siddhi.distribution.core.internal.exception.SiddhiAppDeploymentException;
 import io.siddhi.distribution.core.internal.util.SiddhiAppProcessorConstants;
-import io.siddhi.distribution.core.model.model.SiddhiAppContent;
-import io.siddhi.distribution.core.model.model.SiddhiAppMetrics;
-import io.siddhi.distribution.core.model.model.SiddhiAppRevision;
-import io.siddhi.distribution.core.model.model.SiddhiAppStatus;
+import io.siddhi.distribution.core.model.SiddhiAppContent;
+import io.siddhi.distribution.core.model.SiddhiAppMetrics;
+import io.siddhi.distribution.core.model.SiddhiAppRevision;
+import io.siddhi.distribution.core.model.SiddhiAppStatus;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.wso2.carbon.analytics.permissions.PermissionProvider;
